pyproject.toml
[tool.poetry]
name = "telegram-chatbot"
version = "0.1.0"
description = ""
authors = ["riguima <richard.alexsander.guima@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_chatbot"}]

[tool.poetry.dependencies]
python = "^3.9"
pyrogram = "^2.0.106"
tgcrypto = "^1.2.5"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
blue = "^0.9.1"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
taskipy = "^1.11.0"
ipython = "^8.14.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue --check --diff . && isort --check --diff ."
test = "pytest -s -x --cov=cray_freelas_bot -vv"
post_test = "coverage html"
